Add e2fsprogs to Dockerfile (thanks to Kyle Sower)
Lately the build started to fail with:

```
Do you want to continue? [Y/n] Y
exec: "mkfs.ext4": executable file not found in $PATH
```

`mkfs.ext4` is installed by the package `e2fsprogs`, which used to be
pulled in via some dependency chain (probably `initramfs-tools-core`),
but that is not longer the case. In any case, it's a hard dependency so
it must be listed explicitly in the Dockerfile.

Thanks to Kyle Sower for reporting it at:
<https://gitlab.com/kalilinux/build-scripts/kali-vm/-/merge_requests/14>