Make the swap area easier to configure
Still default to a 1GB swapfile (renamed from /swapfile to /swap).

However, under the hood it's now possible (and easy) to configure a swap
partition (instead of a swap file), and to configure the size, by
setting variables in image.yaml.

The upside of a swap partition is that it becomes possible to have
hibernation, however I'm not certain it's very relevant for VM use-case.

These settings are not exposed in the build.sh script at the moment, but
anyone can open a MR if they think it would be useful. Maybe something
like `-W file,1GB` or `-W part,4GB` would do.