Make vagrant a image format (rather than a new flag)
Rather than using a new flag, we can make 'vagrant' a new image format.
It fits quite well in the existing code, and it simplifies a bunch of
things.

In effect, this commit is a partial revert of the previous commit.

The script vagrant-out.sh becomes export-vagrant.sh, with quite some
refactoring.

We also update the README to show how to build a series of Vagrant
images.