Rename existing user to vagrant user
Overriding USERPASS in build.sh is not the right way to go. Remember
that it's possible (and convenient) to split the build in two: 1) build
a rootfs and 2) build several images out of this rootfs.

A realistic scenario would be to first create a rootfs, then build eg. a
vagrant-virtuabox image and a OVA virtualbox image out of it. As we can
see, for a vagrant-virtualbox we want the user to be vagrant:vagrant,
however for the OVA virtualbox we want the usual kali:kali user.

What it means is that, in step 1 when we create the rootfs, we want to
honor the USERPASS setting (kali:kali by default, or whatever user
passed to build.sh). And it's only in step 2, when we actually create
the vagrant image, that we want to have a vagrant:vagrant user. Given
that, at this stage, there's already a unprivileged user that exists
(kali:kali by default), then let's just rename this existing user.

Additionally, rename the script